parser_backend/requirements.txt
supabase
python-dotenv
python-jose[cryptography]
pdfplumber
pypdf
PyPDF2
google-genai
bcrypt
pandas
pymupdf
camelot-py[base]
opencv-python
fastapi
uvicorn[standard]
httpx
email-validator
python-multipart
requests
